requirements.txt
python-telegram-bot==13.15
web3==5.31.0
openai==0.25.0
numpy==1.19.4